Improved armor values in genstore infos, adjusted vest prices
Instead of a single "Armor" value, there are now "Ballistic Armor" and
"Explosive Armor", in similar fashion to Virtual Arsenal indicators. The
ballistic values are computed from a formula I devised, because the
Arsenal values for ballistic protection are inaccurate. The "Explosive
Armor" should match what was previously displayed as "Armor".

Also, the vest prices are now adjusted relative to the armor values,
with capacity being a lesser factor, so vests with no armor have lower
prices, as it should be.

Thanks to AgentRev